test(taos): modify perf test case
