+++ taos/benches/sql_native_interlace.rs
@@ -112,9 +112,9 @@ async fn produce_sqls(subtable_cnt: usize, record_cnt: usize) -> Vec<String> {
                     for k in 0..batch_cnt {
                         let tbname = format!("d{}", j + k);
                         let c1 = rng.gen::<i32>();
-                        let c2: f32 = rng.gen_range(0.0..1000_0000.);
+                        let c2: f32 = rng.gen_range(0.0..10000000.);
                         let c2 = (c2 * 100.0).round() / 100.0;
-                        let c3: f32 = rng.gen_range(0.0..1000_0000.);
+                        let c3: f32 = rng.gen_range(0.0..10000000.);
                         let c3 = (c3 * 100.0).round() / 100.0;
                         sql.push_str(&format!("{tbname} values({ts},{c1},{c2},{c3}) "));
                     }
