+++ taos/benches/stmt_interlace.rs
@@ -123,9 +123,9 @@ async fn produce_data(
 
                     for k in 0..batch_cnt {
                         let c1 = rng.gen::<i32>();
-                        let c2: f32 = rng.gen_range(0.0..1000_0000.);
+                        let c2: f32 = rng.gen_range(0.0..10000000.);
                         let c2 = (c2 * 100.0).round() / 100.0;
-                        let c3: f32 = rng.gen_range(0.0..1000_0000.);
+                        let c3: f32 = rng.gen_range(0.0..10000000.);
                         let c3 = (c3 * 100.0).round() / 100.0;
 
                         let tbname = format!("d{}", j + k);
@@ -177,7 +177,7 @@ async fn consume_data(db: &str, mut receivers: Vec<Receiver<Vec<(String, Vec<Col
             println!("Consumer thread[{i}] starts consuming data");
 
             let start = Instant::now();
-            while let Ok(datas) = receiver.recv() {
+            while let Ok(datas) = receiver.recv_async().await {
                 for (tbname, cols) in datas {
                     stmt.set_tbname(&tbname).await.unwrap();
                     stmt.bind(&cols).await.unwrap();
